Remove volume mount from docker-compose and tidy nginx conf
If you ever find yourself mounting volumes from the docker-compose.yml this is bad. Use a named volume if data needs to be shared between containers, or add the file directly to the docker image via the dockerfile.

Unnecessary group of servers defined. Remove. Add www.compsoc.tech to http redirects to https.